requirements.txt
openai
streamlit
tiktoken
python-dotenv
supabase
docling
requests
numpy
